+++ components/keyemulate/api/src/main/kotlin/com/flipperdevices/keyemulate/api/EmulateHelper.kt
@@ -1,22 +1,20 @@
 package com.flipperdevices.keyemulate.api
 
 import com.flipperdevices.bridge.api.manager.FlipperRequestApi
-import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
-import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
 import com.flipperdevices.bridge.service.api.FlipperServiceApi
+import com.flipperdevices.keyemulate.model.EmulateConfig
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.StateFlow
 
 const val SUBGHZ_DEFAULT_TIMEOUT_MS = 500L
+const val INFRARED_DEFAULT_TIMEOUT_MS = 500L
 
 interface EmulateHelper {
-    fun getCurrentEmulatingKey(): StateFlow<FlipperFilePath?>
+    fun getCurrentEmulatingKey(): StateFlow<EmulateConfig?>
     suspend fun startEmulate(
         scope: CoroutineScope,
         serviceApi: FlipperServiceApi,
-        keyType: FlipperKeyType,
-        keyPath: FlipperFilePath,
-        minEmulateTime: Long = 0L
+        config: EmulateConfig
     ): Boolean
 
     suspend fun stopEmulate(